+++ src/Type/Definition/InterfaceType.php
@@ -10,7 +10,6 @@ use GraphQL\Language\AST\InterfaceTypeExtensionNode;
 use GraphQL\Type\Schema;
 use GraphQL\Utils\Utils;
 
-use function array_map;
 use function is_array;
 use function is_callable;
 use function is_string;
@@ -104,12 +103,14 @@ class InterfaceType extends TypeWithFields implements AbstractType, OutputType, 
                 );
             }
 
-            /** @var array<int, InterfaceType> $interfaces */
-            $interfaces = $interfaces === null
-                ? []
-                : array_map([Schema::class, 'resolveType'], $interfaces);
-
-            $this->interfaces = $interfaces;
+            $this->interfaces = [];
+            if (is_array($interfaces)) {
+                foreach ($interfaces as $interface) {
+                    /** @var InterfaceType $interfaceType Ensured to be true during schema validation */
+                    $interfaceType      = Schema::resolveType($interface);
+                    $this->interfaces[] = $interfaceType;
+                }
+            }
         }
 
         return $this->interfaces;